requirements.txt
Flask==3.0.3
mysql-connector-python==9.1.0
python-dotenv
